Initialize conscrypt in OkHttpTest
